+++ src/scraper/dto/scrape-command.dto.ts
@@ -1,7 +1,10 @@
+import { HttpProxy } from "../types/http-proxy.class"
 import { InputTarget } from "../types/input-target.class"
 
 export class ScrapeCommandDto {
   url: string
   targets?: InputTarget[] = []
   csr?: boolean = false
+  headers?: Record<string, string>
+  httpProxy?: HttpProxy
 }