+++ src/scraper/types/http-proxy.class.ts
@@ -1,0 +1,8 @@
+export class HttpProxy {
+  host: string
+  port: number
+  auth: {
+    username: string
+    password: string
+  }
+}